+++ programs/locker/src/util/constants.rs
@@ -1,2 +1,3 @@
 pub const TRANSFER_MEMO_CLAIM_VESTING: &str = "Jup-Lock ClaimVesting";
 pub const TRANSFER_MEMO_CANCEL_VESTING: &str = "Jup-Lock CancelVesting";
+pub const TRANSFER_MEMO_CLOSE_ESCROW: &str = "Jup-Lock CloseEscrow";