+++ CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking Changes
 
+## Program [0.3.1] [PR #19]
+
+### Added
+
+- Add new endpoint `close_vesting_escrow` that allows creator of vesting_escrow to close vesting_escrow, escrow_token and escrow_metadata if recipient already claimed all tokens or vesting_escrow is cancelled. 
+
 ## Program [0.3.0] [PR #5](https://github.com/jup-ag/jup-lock/pull/5) [PR #15](https://github.com/jup-ag/jup-lock/pull/15) [PR #18](https://github.com/jup-ag/jup-lock/pull/18)
 
 ### Breaking Changes